adding login and registration page

+++ backend/database/migrations/2022_03_21_122532_create_item_order_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->integer('order_id');
             $table->integer('item_id');
             $table->integer('quantity');
-            $table->float('paid_quantity');
+            $table->decimal('paid_quantity');
             $table->softDeletes();
             $table->timestamps();
         });
